Fix a few bugs *Infinite loop when Audio files are referenced but not included in lift file *Imported chracter set not showing up on first time loading goal *Extraneous code from merge

+++ Backend/Services/LiftService.cs
@@ -263,46 +263,41 @@ namespace BackendFramework.Services
             var importListArr = Directory.GetDirectories(extractedPathToImport);
             var extractedAudioDir = Path.Combine(importListArr.Single(), "audio");
 
-            //add audio
-            foreach (var pro in entry.Pronunciations)
+            if (Directory.Exists(extractedAudioDir))
             {
-                newWord.Audio = pro.Media.FirstOrDefault().Url;
-
-                //get path to ~/AmbigProjName/Import/ExtractedLiftDir/Audio/word mp3
-                var extractedAudioMp3 = Path.Combine(extractedAudioDir, newWord.Audio);
-
-                //move mp3 to audio folder at ~/AmbigProjName/Import/ExtractedLiftDir/Audio/word.mp3
-                var audioFolder = Path.Combine(extractedPathToImport, "Audio");
-                Directory.CreateDirectory(audioFolder);
-                var audioDest = Path.Combine(audioFolder, newWord.Audio);
-                File.Create(audioDest);
-
-                //if there are duplicate filenames then add a (number) like windows does to the end of it
-                int filecount = 1;
-                while (true)
+                //try to read audio b/c it's prob there
+                //add audio
+                foreach (var pro in entry.Pronunciations)
                 {
-                    
-                    try
-                    {
-                        File.Copy(extractedAudioMp3, audioDest);
-                        break; // Exit the loop. Could return from the method, depending
-                               // on what it does...
-                    }
-                    catch(IOException)
+                    newWord.Audio = pro.Media.FirstOrDefault().Url;
+
+                    //get path to ~/AmbigProjName/Import/ExtractedLiftDir/Audio/word mp3
+                    var extractedAudioMp3 = Path.Combine(extractedAudioDir, newWord.Audio);
+
+                    //move mp3 to audio folder at ~/AmbigProjName/Import/ExtractedLiftDir/Audio/word.mp3
+                    var audioFolder = Path.Combine(extractedPathToImport, "Audio");
+                    Directory.CreateDirectory(audioFolder);
+                    var audioDest = Path.Combine(audioFolder, newWord.Audio);
+
+                    //if there are duplicate filenames then add a (number) like windows does to the end of it
+                    int filecount = 1;
+
+                    while (File.Exists(audioDest))
                     {
                         //does the filename already have a counter
                         var filename = Path.GetFileNameWithoutExtension(audioDest);
                         var fileList = Regex.Match(filename, @".+(\([0-9]+\))");
-                        
+
                         //if yes then take it off
-                        if(fileList.Success)
+                        if (fileList.Success)
                         {
                             filename = audioDest.Substring(audioDest.Length - 3);
-                            
+
                         }
                         //add a new counter
-                        audioDest = Path.Combine(audioFolder , Path.GetFileNameWithoutExtension(filename) + "(" + filecount++ + ")" + ".mp3");
+                        audioDest = Path.Combine(audioFolder, Path.GetFileNameWithoutExtension(filename) + "(" + filecount++ + ")" + ".mp3");
                     }
+                    File.Copy(extractedAudioMp3, audioDest);
                 }
             }
